requirements.txt
Flask==3.0.2
tensorflow==2.15.0
numpy
Pillow
gunicorn
